SCons: Move top-level .py files to relevant subfolders
Keep only `version.py` and the general `methods.py` top-level.
