+++ drivers/apple_embedded/SCsub
@@ -1,9 +1,8 @@
 #!/usr/bin/env python
 from misc.utility.scons_hints import *
 
-from SCons.Script import Glob
-
 from platform_methods import setup_swift_builder
+from SCons.Script import Glob
 
 Import("env")
