+++ methods.py
@@ -16,7 +16,6 @@ from pathlib import Path
 from typing import Generator, TextIO, cast
 
 from misc.utility.color import print_error, print_info, print_warning
-from platform_methods import detect_arch
 
 # Get the "Godot" folder name ahead of time
 base_folder = Path(__file__).resolve().parent
@@ -989,527 +988,6 @@ def dump(env):
         file.write(env.Dump(format="json"))
 
 
-# Custom Visual Studio project generation logic that supports any platform that has a msvs.py
-# script, so Visual Studio can be used to run scons for any platform, with the right defines per target.
-# Invoked with scons vsproj=yes
-#
-# Only platforms that opt in to vs proj generation by having a msvs.py file in the platform folder are included.
-# Platforms with a msvs.py file will be added to the solution, but only the current active platform+target+arch
-# will have a build configuration generated, because we only know what the right defines/includes/flags/etc are
-# on the active build target.
-#
-# Platforms that don't support an editor target will have a dummy editor target that won't do anything on build,
-# but will have the files and configuration for the windows editor target.
-#
-# To generate build configuration files for all platforms+targets+arch combinations, users can call
-#   scons vsproj=yes
-# for each combination of platform+target+arch. This will generate the relevant vs project files but
-# skip the build process. This lets project files be quickly generated even if there are build errors.
-#
-# To generate AND build from the command line:
-#   scons vsproj=yes vsproj_gen_only=no
-def generate_vs_project(env, original_args, project_name="godot"):
-    # Augmented glob_recursive that also fills the dirs argument with traversed directories that have content.
-    def glob_recursive_2(pattern, dirs, node="."):
-        from SCons import Node
-        from SCons.Script import Glob
-
-        results = []
-        for f in Glob(str(node) + "/*", source=True):
-            if type(f) is Node.FS.Dir:
-                results += glob_recursive_2(pattern, dirs, f)
-        r = Glob(str(node) + "/" + pattern, source=True)
-        if len(r) > 0 and str(node) not in dirs:
-            d = ""
-            for part in str(node).split("\\"):
-                d += part
-                if d not in dirs:
-                    dirs.append(d)
-                d += "\\"
-        results += r
-        return results
-
-    def get_bool(args, option, default):
-        from SCons.Variables.BoolVariable import _text2bool
-
-        val = args.get(option, default)
-        if val is not None:
-            try:
-                return _text2bool(val)
-            except (ValueError, AttributeError):
-                return default
-        else:
-            return default
-
-    def format_key_value(v):
-        if type(v) in [tuple, list]:
-            return v[0] if len(v) == 1 else f"{v[0]}={v[1]}"
-        return v
-
-    def get_dependencies(file, env, exts, headers, sources, others):
-        for child in file.children():
-            if isinstance(child, str):
-                child = env.File(x)
-            fname = ""
-            try:
-                fname = child.path
-            except AttributeError:
-                # It's not a file.
-                pass
-
-            if fname:
-                parts = os.path.splitext(fname)
-                if len(parts) > 1:
-                    ext = parts[1].lower()
-                    if ext in exts["sources"]:
-                        sources += [fname]
-                    elif ext in exts["headers"]:
-                        headers += [fname]
-                    elif ext in exts["others"]:
-                        others += [fname]
-
-            get_dependencies(child, env, exts, headers, sources, others)
-
-    filtered_args = original_args.copy()
-
-    # Ignore the "vsproj" option to not regenerate the VS project on every build
-    filtered_args.pop("vsproj", None)
-
-    # This flag allows users to regenerate the proj files but skip the building process.
-    # This lets projects be regenerated even if there are build errors.
-    filtered_args.pop("vsproj_gen_only", None)
-
-    # This flag allows users to regenerate only the props file without touching the sln or vcxproj files.
-    # This preserves any customizations users have done to the solution, while still updating the file list
-    # and build commands.
-    filtered_args.pop("vsproj_props_only", None)
-
-    # The "progress" option is ignored as the current compilation progress indication doesn't work in VS
-    filtered_args.pop("progress", None)
-
-    # We add these three manually because they might not be explicitly passed in, and it's important to always set them.
-    filtered_args.pop("platform", None)
-    filtered_args.pop("target", None)
-    filtered_args.pop("arch", None)
-
-    platform = env["platform"]
-    target = env["target"]
-    arch = env["arch"]
-    host_arch = detect_arch()
-
-    host_platform = "windows"
-    if (
-        sys.platform.startswith("linux")
-        or sys.platform.startswith("dragonfly")
-        or sys.platform.startswith("freebsd")
-        or sys.platform.startswith("netbsd")
-        or sys.platform.startswith("openbsd")
-    ):
-        host_platform = "linuxbsd"
-    elif sys.platform == "darwin":
-        host_platform = "macos"
-
-    vs_configuration = {}
-    host_vs_configuration = {}
-    common_build_prefix = []
-    confs = []
-    for x in sorted(glob.glob("platform/*")):
-        # Only platforms that opt in to vs proj generation are included.
-        if not os.path.isdir(x) or not os.path.exists(x + "/msvs.py"):
-            continue
-        tmppath = "./" + x
-        sys.path.insert(0, tmppath)
-        import msvs
-
-        vs_plats = []
-        vs_confs = []
-        try:
-            platform_name = x[9:]
-            vs_plats = msvs.get_platforms()
-            vs_confs = msvs.get_configurations()
-            val = []
-            for plat in vs_plats:
-                val += [{"platform": plat[0], "architecture": plat[1]}]
-
-            vsconf = {"platform": platform_name, "targets": vs_confs, "arches": val}
-            confs += [vsconf]
-
-            # Save additional information about the configuration for the actively selected platform,
-            # so we can generate the platform-specific props file with all the build commands/defines/etc
-            if platform == platform_name:
-                common_build_prefix = msvs.get_build_prefix(env)
-                vs_configuration = vsconf
-            if platform_name == host_platform:
-                host_vs_configuration = vsconf
-                for a in vsconf["arches"]:
-                    if host_arch == a["architecture"]:
-                        host_arch = a["platform"]
-                        break
-        except Exception:
-            pass
-
-        sys.path.remove(tmppath)
-        sys.modules.pop("msvs")
-
-    extensions = {}
-    extensions["headers"] = [".h", ".hh", ".hpp", ".hxx", ".inc", ".inl"]
-    extensions["sources"] = [".c", ".cc", ".cpp", ".cxx", ".m", ".mm", ".java"]
-    extensions["others"] = [".natvis", ".glsl", ".rc"]
-
-    headers = []
-    headers_dirs = []
-    for ext in extensions["headers"]:
-        for file in glob_recursive_2("*" + ext, headers_dirs):
-            headers.append(str(file).replace("/", "\\"))
-
-    sources = []
-    sources_dirs = []
-    for ext in extensions["sources"]:
-        for file in glob_recursive_2("*" + ext, sources_dirs):
-            sources.append(str(file).replace("/", "\\"))
-
-    others = []
-    others_dirs = []
-    for ext in extensions["others"]:
-        for file in glob_recursive_2("*" + ext, others_dirs):
-            others.append(str(file).replace("/", "\\"))
-
-    skip_filters = False
-    import hashlib
-    import json
-
-    md5 = hashlib.md5(
-        json.dumps(sorted(headers + headers_dirs + sources + sources_dirs + others + others_dirs)).encode("utf-8")
-    ).hexdigest()
-
-    if os.path.exists(f"{project_name}.vcxproj.filters"):
-        with open(f"{project_name}.vcxproj.filters", "r", encoding="utf-8") as file:
-            existing_filters = file.read()
-        match = re.search(r"(?ms)^<!-- CHECKSUM$.([0-9a-f]{32})", existing_filters)
-        if match is not None and md5 == match.group(1):
-            skip_filters = True
-
-    import uuid
-
-    # Don't regenerate the filters file if nothing has changed, so we keep the existing UUIDs.
-    if not skip_filters:
-        print(f"Regenerating {project_name}.vcxproj.filters")
-
-        with open("misc/msvs/vcxproj.filters.template", "r", encoding="utf-8") as file:
-            filters_template = file.read()
-        for i in range(1, 10):
-            filters_template = filters_template.replace(f"%%UUID{i}%%", str(uuid.uuid4()))
-
-        filters = ""
-
-        for d in headers_dirs:
-            filters += f'<Filter Include="Header Files\\{d}"><UniqueIdentifier>{{{str(uuid.uuid4())}}}</UniqueIdentifier></Filter>\n'
-        for d in sources_dirs:
-            filters += f'<Filter Include="Source Files\\{d}"><UniqueIdentifier>{{{str(uuid.uuid4())}}}</UniqueIdentifier></Filter>\n'
-        for d in others_dirs:
-            filters += f'<Filter Include="Other Files\\{d}"><UniqueIdentifier>{{{str(uuid.uuid4())}}}</UniqueIdentifier></Filter>\n'
-
-        filters_template = filters_template.replace("%%FILTERS%%", filters)
-
-        filters = ""
-        for file in headers:
-            filters += (
-                f'<ClInclude Include="{file}"><Filter>Header Files\\{os.path.dirname(file)}</Filter></ClInclude>\n'
-            )
-        filters_template = filters_template.replace("%%INCLUDES%%", filters)
-
-        filters = ""
-        for file in sources:
-            filters += (
-                f'<ClCompile Include="{file}"><Filter>Source Files\\{os.path.dirname(file)}</Filter></ClCompile>\n'
-            )
-
-        filters_template = filters_template.replace("%%COMPILES%%", filters)
-
-        filters = ""
-        for file in others:
-            filters += f'<None Include="{file}"><Filter>Other Files\\{os.path.dirname(file)}</Filter></None>\n'
-        filters_template = filters_template.replace("%%OTHERS%%", filters)
-
-        filters_template = filters_template.replace("%%HASH%%", md5)
-
-        with open(f"{project_name}.vcxproj.filters", "w", encoding="utf-8", newline="\r\n") as f:
-            f.write(filters_template)
-
-    headers_active = []
-    sources_active = []
-    others_active = []
-
-    get_dependencies(
-        env.File(f"#bin/godot{env['PROGSUFFIX']}"), env, extensions, headers_active, sources_active, others_active
-    )
-
-    all_items = []
-    properties = []
-    activeItems = []
-    extraItems = []
-
-    set_headers = set(headers_active)
-    set_sources = set(sources_active)
-    set_others = set(others_active)
-    for file in headers:
-        base_path = os.path.dirname(file).replace("\\", "_")
-        all_items.append(f'<ClInclude Include="{file}">')
-        all_items.append(
-            f"  <ExcludedFromBuild Condition=\"!$(ActiveProjectItemList_{base_path}.Contains(';{file};'))\">true</ExcludedFromBuild>"
-        )
-        all_items.append("</ClInclude>")
-        if file in set_headers:
-            activeItems.append(file)
-
-    for file in sources:
-        base_path = os.path.dirname(file).replace("\\", "_")
-        all_items.append(f'<ClCompile Include="{file}">')
-        all_items.append(
-            f"  <ExcludedFromBuild Condition=\"!$(ActiveProjectItemList_{base_path}.Contains(';{file};'))\">true</ExcludedFromBuild>"
-        )
-        all_items.append("</ClCompile>")
-        if file in set_sources:
-            activeItems.append(file)
-
-    for file in others:
-        base_path = os.path.dirname(file).replace("\\", "_")
-        all_items.append(f'<None Include="{file}">')
-        all_items.append(
-            f"  <ExcludedFromBuild Condition=\"!$(ActiveProjectItemList_{base_path}.Contains(';{file};'))\">true</ExcludedFromBuild>"
-        )
-        all_items.append("</None>")
-        if file in set_others:
-            activeItems.append(file)
-
-    if vs_configuration:
-        vsconf = ""
-        for a in vs_configuration["arches"]:
-            if arch == a["architecture"]:
-                vsconf = f"{target}|{a['platform']}"
-                break
-
-        condition = "'$(GodotConfiguration)|$(GodotPlatform)'=='" + vsconf + "'"
-        itemlist = {}
-        for item in activeItems:
-            key = os.path.dirname(item).replace("\\", "_")
-            if key not in itemlist:
-                itemlist[key] = [item]
-            else:
-                itemlist[key] += [item]
-
-        for x in itemlist.keys():
-            properties.append(
-                "<ActiveProjectItemList_%s>;%s;</ActiveProjectItemList_%s>" % (x, ";".join(itemlist[x]), x)
-            )
-        output = os.path.join("bin", f"godot{env['PROGSUFFIX']}")
-
-        # The modules_enabled.gen.h header containing the defines is only generated on build, and only for the most recently built
-        # platform, which means VS can't properly render code that's inside module-specific ifdefs. This adds those defines to the
-        # platform-specific VS props file, so that VS knows which defines are enabled for the selected platform.
-        env.Append(VSHINT_DEFINES=[f"MODULE_{module.upper()}_ENABLED" for module in env.module_list])
-
-        with open("misc/msvs/props.template", "r", encoding="utf-8") as file:
-            props_template = file.read()
-
-        props_template = props_template.replace("%%CONDITION%%", condition)
-        props_template = props_template.replace("%%PROPERTIES%%", "\n    ".join(properties))
-        props_template = props_template.replace("%%EXTRA_ITEMS%%", "\n    ".join(extraItems))
-
-        props_template = props_template.replace("%%OUTPUT%%", output)
-
-        proplist = [format_key_value(j) for j in list(env["CPPDEFINES"])]
-        proplist += [format_key_value(j) for j in env.get("VSHINT_DEFINES", [])]
-        props_template = props_template.replace("%%DEFINES%%", ";".join(proplist))
-
-        proplist = [str(j) for j in env["CPPPATH"]]
-        proplist += [str(j) for j in env.get("VSHINT_INCLUDES", [])]
-        proplist += [str(j) for j in get_default_include_paths(env)]
-        props_template = props_template.replace("%%INCLUDES%%", ";".join(proplist))
-
-        proplist = [env.subst("$CCFLAGS")]
-        proplist += [env.subst("$CXXFLAGS")]
-        proplist += [env.subst("$VSHINT_OPTIONS")]
-        props_template = props_template.replace("%%OPTIONS%%", " ".join(proplist))
-
-        # Windows allows us to have spaces in paths, so we need
-        # to double quote off the directory. However, the path ends
-        # in a backslash, so we need to remove this, lest it escape the
-        # last double quote off, confusing MSBuild
-        common_build_postfix = [
-            "--directory=&quot;$(ProjectDir.TrimEnd(&apos;\\&apos;))&quot;",
-            "progress=no",
-            f"platform={platform}",
-            f"target={target}",
-            f"arch={arch}",
-        ]
-
-        for arg, value in filtered_args.items():
-            common_build_postfix.append(f"{arg}={value}")
-
-        cmd_rebuild = [
-            "vsproj=yes",
-            "vsproj_props_only=yes",
-            "vsproj_gen_only=no",
-            f"vsproj_name={project_name}",
-        ] + common_build_postfix
-
-        cmd_clean = [
-            "--clean",
-        ] + common_build_postfix
-
-        commands = "scons"
-        if len(common_build_prefix) == 0:
-            commands = "echo Starting SCons &amp; " + commands
-        else:
-            common_build_prefix[0] = "echo Starting SCons &amp; " + common_build_prefix[0]
-
-        cmd = " ".join(common_build_prefix + [" ".join([commands] + common_build_postfix)])
-        props_template = props_template.replace("%%BUILD%%", cmd)
-
-        cmd = " ".join(common_build_prefix + [" ".join([commands] + cmd_rebuild)])
-        props_template = props_template.replace("%%REBUILD%%", cmd)
-
-        cmd = " ".join(common_build_prefix + [" ".join([commands] + cmd_clean)])
-        props_template = props_template.replace("%%CLEAN%%", cmd)
-
-        with open(
-            f"{project_name}.{platform}.{target}.{arch}.generated.props", "w", encoding="utf-8", newline="\r\n"
-        ) as f:
-            f.write(props_template)
-
-    proj_uuid = str(uuid.uuid4())
-    sln_uuid = str(uuid.uuid4())
-
-    if os.path.exists(f"{project_name}.sln"):
-        for line in open(f"{project_name}.sln", "r", encoding="utf-8").read().splitlines():
-            if line.startswith('Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}")'):
-                proj_uuid = re.search(
-                    r"\"{(\b[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-\b[0-9a-fA-F]{12}\b)}\"$",
-                    line,
-                ).group(1)
-            elif line.strip().startswith("SolutionGuid ="):
-                sln_uuid = re.search(
-                    r"{(\b[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-\b[0-9a-fA-F]{12}\b)}", line
-                ).group(1)
-                break
-
-    configurations = []
-    imports = []
-    properties = []
-    section1 = []
-    section2 = []
-    for conf in confs:
-        godot_platform = conf["platform"]
-        has_editor = "editor" in conf["targets"]
-
-        # Skip any platforms that can build the editor and don't match the host platform.
-        #
-        # When both Windows and Mac define an editor target, it's defined as platform+target+arch (windows+editor+x64 for example).
-        # VS only supports two attributes, a "Configuration" and a "Platform", and we currently map our target to the Configuration
-        # (i.e. editor/template_debug/template_release), and our architecture to the "Platform" (i.e. x64, arm64, etc).
-        # Those two are not enough to disambiguate multiple godot targets for different godot platforms with the same architecture,
-        # i.e. editor|x64 would currently match both windows editor intel 64 and linux editor intel 64.
-        #
-        # TODO: More work is needed in order to support generating VS projects that unambiguously support all platform+target+arch variations.
-        # The VS "Platform" has to be a known architecture that VS recognizes, so we can only play around with the "Configuration" part of the combo.
-        if has_editor and godot_platform != host_vs_configuration["platform"]:
-            continue
-
-        for p in conf["arches"]:
-            sln_plat = p["platform"]
-            proj_plat = sln_plat
-            godot_arch = p["architecture"]
-
-            # Redirect editor configurations for platforms that don't support the editor target to the default editor target on the
-            # active host platform, so the solution has all the permutations and VS doesn't complain about missing project configurations.
-            # These configurations are disabled, so they show up but won't build.
-            if not has_editor:
-                section1 += [f"editor|{sln_plat} = editor|{proj_plat}"]
-                section2 += [f"{{{proj_uuid}}}.editor|{proj_plat}.ActiveCfg = editor|{host_arch}"]
-
-                configurations += [
-                    f'<ProjectConfiguration Include="editor|{proj_plat}">',
-                    "  <Configuration>editor</Configuration>",
-                    f"  <Platform>{proj_plat}</Platform>",
-                    "</ProjectConfiguration>",
-                ]
-
-                properties += [
-                    f"<PropertyGroup Condition=\"'$(Configuration)|$(Platform)'=='editor|{proj_plat}'\">",
-                    "  <GodotConfiguration>editor</GodotConfiguration>",
-                    f"  <GodotPlatform>{proj_plat}</GodotPlatform>",
-                    "</PropertyGroup>",
-                ]
-
-            for t in conf["targets"]:
-                godot_target = t
-
-                # Windows x86 is a special little flower that requires a project platform == Win32 but a solution platform == x86.
-                if godot_platform == "windows" and godot_target == "editor" and godot_arch == "x86_32":
-                    sln_plat = "x86"
-
-                configurations += [
-                    f'<ProjectConfiguration Include="{godot_target}|{proj_plat}">',
-                    f"  <Configuration>{godot_target}</Configuration>",
-                    f"  <Platform>{proj_plat}</Platform>",
-                    "</ProjectConfiguration>",
-                ]
-
-                properties += [
-                    f"<PropertyGroup Condition=\"'$(Configuration)|$(Platform)'=='{godot_target}|{proj_plat}'\">",
-                    f"  <GodotConfiguration>{godot_target}</GodotConfiguration>",
-                    f"  <GodotPlatform>{proj_plat}</GodotPlatform>",
-                    "</PropertyGroup>",
-                ]
-
-                p = f"{project_name}.{godot_platform}.{godot_target}.{godot_arch}.generated.props"
-                imports += [
-                    f'<Import Project="$(MSBuildProjectDirectory)\\{p}" Condition="Exists(\'$(MSBuildProjectDirectory)\\{p}\')"/>'
-                ]
-
-                section1 += [f"{godot_target}|{sln_plat} = {godot_target}|{sln_plat}"]
-
-                section2 += [
-                    f"{{{proj_uuid}}}.{godot_target}|{sln_plat}.ActiveCfg = {godot_target}|{proj_plat}",
-                    f"{{{proj_uuid}}}.{godot_target}|{sln_plat}.Build.0 = {godot_target}|{proj_plat}",
-                ]
-
-    # Add an extra import for a local user props file at the end, so users can add more overrides.
-    imports += [
-        f'<Import Project="$(MSBuildProjectDirectory)\\{project_name}.vs.user.props" Condition="Exists(\'$(MSBuildProjectDirectory)\\{project_name}.vs.user.props\')"/>'
-    ]
-    section1 = sorted(section1)
-    section2 = sorted(section2)
-
-    if not get_bool(original_args, "vsproj_props_only", False):
-        with open("misc/msvs/vcxproj.template", "r", encoding="utf-8") as file:
-            proj_template = file.read()
-        proj_template = proj_template.replace("%%UUID%%", proj_uuid)
-        proj_template = proj_template.replace("%%CONFS%%", "\n    ".join(configurations))
-        proj_template = proj_template.replace("%%IMPORTS%%", "\n  ".join(imports))
-        proj_template = proj_template.replace("%%DEFAULT_ITEMS%%", "\n    ".join(all_items))
-        proj_template = proj_template.replace("%%PROPERTIES%%", "\n  ".join(properties))
-
-        with open(f"{project_name}.vcxproj", "w", encoding="utf-8", newline="\r\n") as f:
-            f.write(proj_template)
-
-    if not get_bool(original_args, "vsproj_props_only", False):
-        with open("misc/msvs/sln.template", "r", encoding="utf-8") as file:
-            sln_template = file.read()
-        sln_template = sln_template.replace("%%NAME%%", project_name)
-        sln_template = sln_template.replace("%%UUID%%", proj_uuid)
-        sln_template = sln_template.replace("%%SLNUUID%%", sln_uuid)
-        sln_template = sln_template.replace("%%SECTION1%%", "\n\t\t".join(section1))
-        sln_template = sln_template.replace("%%SECTION2%%", "\n\t\t".join(section2))
-
-        with open(f"{project_name}.sln", "w", encoding="utf-8", newline="\r\n") as f:
-            f.write(sln_template)
-
-    if get_bool(original_args, "vsproj_gen_only", True):
-        sys.exit()
-
-
 ############################################################
 # FILE GENERATION & FORMATTING
 ############################################################
@@ -1671,18 +1149,3 @@ def to_raw_cstring(value: str | list[str]) -> str:
     else:
         # Wrap multiple segments in parenthesis to suppress `string-concatenation` warnings on clang.
         return "({})".format(" ".join(f'R"<!>({segment.decode()})<!>"' for segment in split))
-
-
-def get_default_include_paths(env):
-    if env.msvc:
-        return []
-    compiler = env.subst("$CXX")
-    target = os.path.join(env.Dir("#main").abspath, "main.cpp")
-    args = [compiler, target, "-x", "c++", "-v"]
-    ret = subprocess.run(args, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-    output = ret.stdout
-    match = re.search(r"#include <\.\.\.> search starts here:([\S\s]*)End of search list.", output)
-    if not match:
-        print_warning("Failed to find the include paths in the compiler output.")
-        return []
-    return [x.strip() for x in match[1].strip().splitlines()]
