+++ misc/utility/scu_builders.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 from methods import print_error
 
-base_folder_path = str(Path(__file__).parent) + "/"
+base_folder_path = str(Path(__file__).parent.parent.parent) + "/"
 base_folder_only = os.path.basename(os.path.normpath(base_folder_path))
 _verbose = False  # Set manually for debug prints
 _scu_folders = set()
