+++ modules/betsy/SCsub
@@ -13,7 +13,7 @@ env_betsy.GLSL_HEADER("bc4.glsl")
 env_betsy.GLSL_HEADER("alpha_stitch.glsl")
 env_betsy.GLSL_HEADER("rgb_to_rgba.glsl")
 
-env_betsy.Depends(Glob("*.glsl.gen.h"), ["#glsl_builders.py"])
+env_betsy.Depends(Glob("*.glsl.gen.h"), ["#servers/rendering/glsl_builders.py"])
 
 # Godot source files
 env_betsy.add_source_files(env.modules_sources, "*.cpp")
