+++ modules/lightmapper_rd/SCsub
@@ -8,7 +8,7 @@ env_lightmapper_rd = env_modules.Clone()
 env_lightmapper_rd.GLSL_HEADER("lm_raster.glsl")
 env_lightmapper_rd.GLSL_HEADER("lm_compute.glsl")
 env_lightmapper_rd.GLSL_HEADER("lm_blendseams.glsl")
-env_lightmapper_rd.Depends(Glob("*.glsl.gen.h"), ["lm_common_inc.glsl", "#glsl_builders.py"])
+env_lightmapper_rd.Depends(Glob("*.glsl.gen.h"), ["lm_common_inc.glsl", "#servers/rendering/glsl_builders.py"])
 
 # Godot source files
 env_lightmapper_rd.add_source_files(env.modules_sources, "*.cpp")
