+++ platform/ios/detect.py
@@ -2,8 +2,9 @@ import os
 import sys
 from typing import TYPE_CHECKING
 
-from methods import detect_darwin_sdk_path, detect_darwin_toolchain_path, print_error, print_warning
 from platform_methods import validate_arch
+
+from methods import detect_darwin_sdk_path, detect_darwin_toolchain_path, print_error, print_warning
 
 if TYPE_CHECKING:
     from SCons.Script.SConscript import SConsEnvironment
