+++ platform/linuxbsd/detect.py
@@ -3,8 +3,9 @@ import platform
 import sys
 from typing import TYPE_CHECKING
 
-from methods import get_compiler_version, print_error, print_info, print_warning, using_gcc
 from platform_methods import detect_arch, validate_arch
+
+from methods import get_compiler_version, print_error, print_info, print_warning, using_gcc
 
 if TYPE_CHECKING:
     from SCons.Script.SConscript import SConsEnvironment
