+++ platform/visionos/SCsub
@@ -1,9 +1,8 @@
 #!/usr/bin/env python
 from misc.utility.scons_hints import *
 
-from platform_visionos_builders import generate_bundle
-
 from platform_methods import combine_libs_apple_embedded
+from platform_visionos_builders import generate_bundle
 
 Import("env")
