+++ platform/windows/detect.py
@@ -4,9 +4,10 @@ import subprocess
 import sys
 from typing import TYPE_CHECKING
 
+from platform_methods import detect_arch, validate_arch
+
 import methods
 from methods import print_error, print_warning
-from platform_methods import detect_arch, validate_arch
 
 if TYPE_CHECKING:
     from SCons.Script.SConscript import SConsEnvironment
