+++ servers/rendering/renderer_rd/shaders/effects/fsr2/SCsub
@@ -16,7 +16,7 @@ if "RD_GLSL" in env["BUILDERS"]:
     glsl_files = [str(f) for f in Glob("*.glsl") if str(f) not in gl_include_files]
 
     # make sure we recompile shaders if include files change
-    env.Depends([f + ".gen.h" for f in glsl_files], gl_include_files + ["#glsl_builders.py"])
+    env.Depends([f + ".gen.h" for f in glsl_files], gl_include_files + ["#servers/rendering/glsl_builders.py"])
 
     # compile shaders
     for glsl_file in glsl_files:
